Put variable expansion in its own module
